API reference

Public

DifferentiableExpectations.distribution — Function

distribution(e::AbstractExpectation, θ...)

Build the sampling distribution for e based on parameters heta.

DifferentiableExpectations.samples — Method

samples(e::AbstractExpectation, θ...)

Return the values of the function represented by e for several samples of distribution(e, θ...).

Private

DifferentiableExpectations.AbstractExpectation — Type

AbstractExpectation{threaded}

Abstract supertype for expectation wrappers.

The type parameter threaded is a Bool stating whether the Monte-Carlo samples should be computed in parallel.

DifferentiableExpectations.AbstractExpectation — Method

(e::AbstractExpectation)(θ...)

Return the Monte-Carlo average of the function represented by e over several samples of distribution(e, θ...).