API reference

Public

DifferentiableExpectations.DifferentiableExpectations — Module

DifferentiableExpectations

A Julia package for differentiating through expectations with Monte-Carlo estimates.

Exports

DifferentiableExpectation

FixedAtomsProbabilityDistribution

Reinforce

Reparametrization

DifferentiableExpectations.DifferentiableExpectation — Type

DifferentiableExpectation{threaded}

Abstract supertype for differentiable parametric expectations F : θ -> 𝔼[f(X)] where X ∼ p(θ), whose value and derivative are approximated with Monte-Carlo averages.

Subtypes

Reinforce

Reparametrization

Calling behavior

(F::DifferentiableExpectation)(θ...; kwargs...)

Return a Monte-Carlo average (1/s) ∑f(xᵢ) where the xᵢ ∼ p(θ) are iid samples.

Type parameters

threaded::Bool: specifies whether the sampling should be performed in parallel

Required fields

f: The function applied inside the expectation.

dist_constructor: The constructor of the probability distribution.

rng::AbstractRNG: The random number generator.

nb_samples::Integer: The number of Monte-Carlo samples.

The field dist_constructor must be a callable such that dist_constructor(θ...) generates an object dist that corresponds to p(θ). The resulting object dist needs to satisfy:

the Random API for sampling with rand(rng, dist)

the DensityInterface.jl API for loglikelihoods with logdensityof(dist, x)

DifferentiableExpectations.FixedAtomsProbabilityDistribution — Type

FixedAtomsProbabilityDistribution{threaded}

A probability distribution with finite support and fixed atoms.

Whenever its expectation is differentiated, only the weights are considered active, whereas the atoms are considered constant.

Example

julia> using DifferentiableExpectations, Statistics, Zygote

julia> dist = FixedAtomsProbabilityDistribution([2, 3], [0.4, 0.6]);

julia> map(abs2, dist)
FixedAtomsProbabilityDistribution{false}([4, 9], [0.4, 0.6])

julia> mean(abs2, dist)
7.0

julia> gradient(mean, abs2, dist)[2]
(atoms = nothing, weights = [4.0, 9.0])

Constructor

FixedAtomsProbabilityDistribution(
    atoms::Vector,
    weights::Vector;
    threaded=false
)

Fields

atoms::Vector

weights::Vector{W} where W<:Real

DifferentiableExpectations.Reinforce — Type

Reinforce{threaded} <: DifferentiableExpectation{threaded}

Differentiable parametric expectation F : θ -> 𝔼[f(X)] where X ∼ p(θ) using the REINFORCE (or score function) gradient estimator:

∂F(θ) = 𝔼[f(X) ∇₂logp(X,θ)ᵀ]

Example

using DifferentiableExpectations, Distributions, Zygote

F = Reinforce(exp, Normal; nb_samples=10^5)
F_true(μ, σ) = mean(LogNormal(μ, σ))

μ, σ = 0.5, 1,0
∇F, ∇F_true = gradient(F, μ, σ), gradient(F_true, μ, σ)
isapprox(collect(∇F), collect(∇F_true); rtol=1e-1)

# output

true

Constructor

Reinforce(
    f,
    dist_constructor,
    dist_logdensity_grad=nothing;
    rng=Random.default_rng(),
    nb_samples=1,
    threaded=false
)

Fields

f::Any: function applied inside the expectation

dist_constructor::Any: constructor of the probability distribution (θ...) -> p(θ)

dist_logdensity_grad::Any: either nothing or a parameter gradient callable (x, θ...) -> ∇₂logp(x, θ)

rng::Random.AbstractRNG: random number generator

nb_samples::Int64: number of Monte-Carlo samples

See also

DifferentiableExpectation

DifferentiableExpectations.Reparametrization — Type

Reparametrization{threaded} <: DifferentiableExpectation{threaded}

Differentiable parametric expectation F : θ -> 𝔼[f(X)] where X ∼ p(θ) using the reparametrization (or pathwise) gradient estimator: if X = g(Z,θ) where Z ∼ q then

∂F(θ) = 𝔼_q[∂f(g(Z,θ)) ∂₂g(Z,θ)ᵀ]

Example

using DifferentiableExpectations, Distributions, Zygote

F = Reparametrization(exp, Normal; nb_samples=10^4)
F_true(μ, σ) = mean(LogNormal(μ, σ))

μ, σ = 0.5, 1,0
∇F, ∇F_true = gradient(F, μ, σ), gradient(F_true, μ, σ)
isapprox(collect(∇F), collect(∇F_true); rtol=1e-1)

# output

true

Constructor

Reparametrization(
    f,
    dist_constructor,
    rng=Random.default_rng(),
    nb_samples=1,
    threaded=false
)

Fields

f::Any: function applied inside the expectation

dist_constructor::Any: constructor of the probability distribution (θ...) -> p(θ)

rng::Random.AbstractRNG: random number generator

nb_samples::Int64: number of Monte-Carlo samples

See also

DifferentiableExpectation

Private

DifferentiableExpectations.TransformedDistribution — Type

TransformedDistribution

Represent the probability distribution p of a random variable X ∼ p with a transformation X = T(Z) where Z ∼ q.

Fields

base_dist::Any: the distribution q that gets transformed into p

transformation::Any: the transformation function T

Base.map — Method

map(f, dist::FixedAtomsProbabilityDistribution)

Apply f to the atoms of dist, leave the weights unchanged.

Base.rand — Method

rand(rng, dist::TransformedDistribution)

Sample from dist by applying dist.transformation to dist.base_dist.

Base.rand — Method

rand(rng, dist::FixedAtomsProbabilityDistribution)

Sample from the atoms of dist with probability proportional to their weights.

DifferentiableExpectations.presamples — Method

presamples(F::DifferentiableExpectation, θ...)

Return a vector [x₁, ..., xₛ] or matrix [x₁ ... xₛ] where the xᵢ ∼ p(θ) are iid samples.

DifferentiableExpectations.reparametrize — Function

reparametrize(dist)

Turn a probability distribution p into a TransformedDistribution (q, T) such that the new distribution q does not depend on the parameters of p. These parameters are encoded (closed over) in the transformation function T.

DifferentiableExpectations.samples — Method

samples(F::DifferentiableExpectation, θ...; kwargs...)

Return a vector [f(x₁), ..., f(xₛ)] where the xᵢ ∼ p(θ) are iid samples.

Statistics.mean — Method

mean(f, dist::FixedAtomsProbabilityDistribution)

Shortcut for mean(map(f, dist)).

Statistics.mean — Method

mean(dist::FixedAtomsProbabilityDistribution)

Compute the expectation of dist, i.e. the sum of all atoms multiplied by their respective weights.